requirements.txt
alembic==1.0.5
Click==7.0
decorator==4.3.0
dicttoxml==1.7.4
Flask==1.0.2
Flask-CKEditor==0.4.3
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-Pagination==0.0.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
pbr==5.1.1
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
Werkzeug==0.14.1
WTForms==2.2.1
